+++ Heroes.java
@@ -1,6 +1,6 @@
 public class Heroes extends Personaje{
-    private static int ataqueHeroes=100;
-    public Heroes(int vida, String nombre, int armadura){
+    private static double ataqueHeroes=100;
+    public Heroes(double vida, String nombre, double armadura){
         super(vida,nombre,ataqueHeroes,armadura);
     }
